+++ hypertools/io/__init__.py
@@ -2,6 +2,6 @@ from .load import load
 from .save import save
 from . import sources
 from . import streaming
-from .lsl import lsl_stream
+from .lsl import lsl_stream, LSLStream
 
-__all__ = ['load', 'save', 'sources', 'streaming', 'lsl_stream']
+__all__ = ['load', 'save', 'sources', 'streaming', 'lsl_stream', 'LSLStream']